Refactor skill loading logic in integration tests
      Summary: This commit extracts the common logic for loading skills from directories into a TestHelpers class, and uses it in the OpenAICompletionTests and PlannerSkillTests classes. This reduces code duplication and improves readability. The commit also adds a method to import all the sample skills at once, which can be useful for testing scenarios that require multiple skills.
